% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{calculateMindist}
\alias{calculateMindist}
\alias{calculateMindist,TrioSet-method}
\alias{calculateMindist,TrioSetList-method}
\alias{calculateMindist,arrayORff_array-method}
\alias{calculateMindist,list-method}
\alias{calculateMindist,matrix-method}
\title{Deprecated function to calculate the minimum distance}
\usage{
calculateMindist(object, ...)

\S4method{calculateMindist}{TrioSet}(object, verbose = TRUE, ...)

\S4method{calculateMindist}{TrioSetList}(object)

\S4method{calculateMindist}{list}(object, outdir = ldPath(), ...)

\S4method{calculateMindist}{arrayORff_array}(object, outdir, center, ...)

\S4method{calculateMindist}{matrix}(object, ...)
}
\arguments{
\item{object}{see \code{showMethods("calculateMindist")}}

\item{...}{Ignored}

\item{verbose}{logical. Whether to display messages indicating progress.}

\item{outdir}{character string indicating path to save output}

\item{center}{ignored}
}
\description{
The 'minimum distance' is the minimum signed absolute difference of
   the parental log R ratios and the offspring log R
   ratios. Specifically, let |O-F| denote the absolute difference in
   the log R ratios comparing offspring to father and |O-M| the
   absolute difference in the log R ratios comparing offspring to
   mother.  The minimum distance at a marker is the signed minimum of
   |O-M| and |O-F|.  After segmentation of the minimum distance,
   non-zero segments can indicate a de novo difference in the log R
   ratio of the offspring and either parent.  For example, a positive
   minimum distance suggests that the log R ratio from the offspring is
   greater than the log R ratio of either parent.
}

