% Generated by roxygen2 (4.0.2): do not edit by hand
\name{offspringNames}
\alias{offspringNames}
\alias{offspringNames<-}
\alias{pedigree}
\alias{pedigree<-}
\alias{trios}
\title{Deprecated functions and methods}
\usage{
offspringNames(object)

offspringNames(object) <- value

trios(object)

pedigree(object)

pedigree(object) <- value
}
\arguments{
\item{object}{see \code{showMethods("offspringNames")}}

\item{value}{a character vector of offspring identifiers}
}
\description{
These functions will be defunct in a future release.
}

