% Generated by roxygen2 (4.0.2): do not edit by hand
\name{DNAcopyParam}
\alias{DNAcopyParam}
\title{Constructor for DNAcopyParam class}
\usage{
DNAcopyParam(alpha = 0.01, min.width = 2L, undo.splits = c("none",
  "prune", "sdundo"), undo.SD = 3)
}
\arguments{
\item{alpha}{see \code{\link[DNAcopy]{segment}}}

\item{min.width}{see \code{\link[DNAcopy]{segment}}}

\item{undo.splits}{see \code{\link[DNAcopy]{segment}}}

\item{undo.SD}{see \code{\link[DNAcopy]{segment}}}
}
\description{
Creates an instance of a parameter class for circular binary
segmentation of the minimum distance and the log R ratios.
Parameters in this object are passed to the \code{segment} function
in the package DNAcopy.
}
\examples{
segment_params <- DNAcopyParam(alpha=0.01)
params <- MinDistParam(dnacopy=segment_params)
}
\seealso{
\code{\link[DNAcopy]{segment}}
}

