% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{filterExperiment}
\alias{filterExperiment}
\alias{filterExperiment,MinDistExperiment,GRanges-method}
\alias{filterExperiment,MinDistExperiment,GRangesList-method}
\alias{filterExperiment,MinDistExperiment,MinDistGRanges-method}
\title{Methods for filtering MinDistExperiment objects}
\usage{
filterExperiment(object, granges, param)

\S4method{filterExperiment}{MinDistExperiment,GRanges}(object, granges, param)

\S4method{filterExperiment}{MinDistExperiment,GRangesList}(object, granges,
  param)

\S4method{filterExperiment}{MinDistExperiment,MinDistGRanges}(object, granges,
  param)
}
\arguments{
\item{object}{A \code{MinDistExperiment}}

\item{granges}{A \code{GRanges}, \code{GRangesList}, or \code{MinDistGRanges} object}

\item{param}{a \code{MinDistParam} object}
}
\value{
a \code{MinDistExperiment}
}
\description{
Filter a MinDistExperiment object to exclude markers with missing
values in the low-level summaries, exclude markers that lie in
segments (\code{granges} argument) with small minimum distance
values (unlikely to be de novo)
}

