% Generated by roxygen2 (4.0.2): do not edit by hand
\name{md_exp}
\alias{md_exp}
\title{An example \code{MinDistExperiment}}
\format{a \code{MinDistExperiment}}
\description{
This dataset contains log R ratios and B allele frequencies from a
parent-offspring trio (three individuals). Only markers from
chromosomes 7 and 22 are included in this object. The
\code{MinDistExperiment} class extends \code{SummarizedExperiment},
and so many of the methods defined for \code{SummarizedExperiment}
such as \code{findOverlaps} are available through inheritance.
}

