% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mdLegend}
\alias{mdLegend}
\title{Text summary of information encapculated in a MDRanges object for a particular interval}
\usage{
mdLegend(g)
}
\arguments{
\item{g}{a \code{MDRanges} object}
}
\description{
Text summary of information encapculated in a MDRanges object for a particular interval
}

