% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mindist}
\alias{mindist}
\alias{mindist<-}
\title{Getter and setter for the minimum distance statistic}
\usage{
mindist(object)

mindist(object) <- value
}
\arguments{
\item{object}{see \code{showMethods("mindist")}}

\item{value}{a matrix of the minimum distance}
}
\description{
Getter and setter for the minimum distance statistic
}

