% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{MinDistExperiment-class}
\alias{MinDistExperiment-class}
\alias{[,MinDistExperiment,ANY,ANY,ANY-method}
\alias{father,MinDistExperiment-method}
\alias{father,SummarizedExperiment-method}
\alias{mindist,MinDistExperiment-method}
\alias{mindist<-,MinDistExperiment,ANY-method}
\alias{mother,MinDistExperiment-method}
\alias{mother,SummarizedExperiment-method}
\alias{offspring,MinDistExperiment-method}
\alias{offspring,SummarizedExperiment-method}
\alias{pedigree,MinDistExperiment-method}
\alias{pedigree<-,MinDistExperiment-method}
\alias{segment2,MinDistExperiment-method}
\alias{show,MinDistExperiment-method}
\title{Class and methods for MinDistExperiment}
\usage{
\S4method{show}{MinDistExperiment}(object)

\S4method{pedigree}{MinDistExperiment}(object)

\S4method{pedigree}{MinDistExperiment}(object) <- value

\S4method{mindist}{MinDistExperiment}(object)

\S4method{mindist}{MinDistExperiment,ANY}(object) <- value

\S4method{[}{MinDistExperiment,ANY,ANY,ANY}(x, i, j, ..., drop = FALSE)

\S4method{offspring}{MinDistExperiment}(object)

\S4method{father}{MinDistExperiment}(object)

\S4method{mother}{MinDistExperiment}(object)

\S4method{segment2}{MinDistExperiment}(object, param = MinDistParam())

\S4method{father}{SummarizedExperiment}(object)

\S4method{mother}{SummarizedExperiment}(object)

\S4method{offspring}{SummarizedExperiment}(object)
}
\arguments{
\item{object}{a \code{MinDistExperiment} object}

\item{value}{a \code{ParentOffspring} object}

\item{x}{a \code{MinDistExperiment} object}

\item{i}{a numeric-vector for indexing the rows (optional)}

\item{j}{a numeric-vector for indexing the columns (optional)}

\item{...}{additional arguments propogated to subsetting methods for \code{SummarizedExperiment}}

\item{drop}{logical. Whether to simplify a one-row or one-column
matrix to a vector. In most cases, this should always be FALSE.}

\item{param}{a \code{MinDistParam} object}
}
\description{
Class and methods for MinDistExperiment
}
\section{Slots}{

\describe{
\item{\code{mindist}}{a matrix}

\item{\code{pedigree}}{a \code{ParentOffspring} object}
}}

