% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{MinDistExperiment}
\alias{MinDistExperiment}
\alias{MinDistExperiment,ArrayViews,ParentOffspring-method}
\title{Constructor for \code{MinDistExperiment} class}
\usage{
MinDistExperiment(object = ArrayViews(), pedigree = ParentOffspring(), ...)


  \S4method{MinDistExperiment}{ArrayViews,ParentOffspring}(object = ArrayViews(),
  pedigree = ParentOffspring(), ...)
}
\arguments{
\item{object}{see \code{showMethods(MinDistExperiment)}}

\item{pedigree}{a \code{ParentOffspring} object}

\item{...}{ignored}
}
\value{
an object of class \code{MinDistExperiment}
}
\description{
Constructor for \code{MinDistExperiment} class
}

