% Generated by roxygen2 (4.0.2): do not edit by hand
\name{MinDistGRanges}
\alias{MinDistGRanges}
\title{Constructor for \code{MinDistGRanges} class}
\usage{
MinDistGRanges(mindist = GRangesList(), offspring = GRangesList(),
  father = GRanges(), mother = GRanges(), pedigree = ParentOffspring())
}
\arguments{
\item{mindist}{a \code{GRangesList} object}

\item{offspring}{a \code{GRangesList} object}

\item{father}{a \code{GRanges} object}

\item{mother}{a \code{GRanges} object}

\item{pedigree}{a \code{ParentOffspring} object}
}
\description{
The \code{MinDistGRanges} class contains the segmentation of the
father, mother, offspring, and the minimum distance for each
possible parent-offspring trio.  For the parents, the segmentation
results are expected to be in \code{GRanges} format.  To accomodate
multiple-offspring families, both the offspring segments and
minimum distance segments should be of class \code{GRangesList}
where the length of the list corresponds to the number of
offspring.
}
\examples{
MinDistGRanges()
}

