% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{MinDistParam-class}
\alias{MinDistParam-class}
\alias{nMAD,MinDistParam-method}
\alias{nMAD<-,MinDistParam,numeric-method}
\alias{show,MinDistParam-method}
\title{Class and methods for parameters of minimum distance algorithm}
\usage{
\S4method{nMAD}{MinDistParam}(object)

\S4method{nMAD}{MinDistParam,numeric}(object) <- value

\S4method{show}{MinDistParam}(object)
}
\arguments{
\item{object}{a \code{MinDistParam} object}

\item{value}{a length-one numeric vector.}
}
\description{
Contains parameters used for circular binary segmentation (package
DNAcopy), parameters in the PennCNV hidden Markov model, and
parameters used for computing emission probabilities.
}
\section{Slots}{

\describe{
\item{\code{nMAD}}{a length-one numeric vector}

\item{\code{dnacopy}}{an object of class \code{DNAcopyParam}}

\item{\code{penncnv}}{an object of class \code{PennParam}}

\item{\code{emission}}{an object of class \code{EmissionParam}}

\item{\code{thin}}{a length-one non-negative integer}
}}

