% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{MinDistPosterior-class}
\alias{MinDistPosterior-class}
\alias{[,MinDistPosterior,ANY,ANY,ANY-method}
\alias{show,MinDistPosterior-method}
\title{Container for the segmentation results from a MinDistExperiment}
\usage{
\S4method{show}{MinDistPosterior}(object)

\S4method{[}{MinDistPosterior,ANY,ANY,ANY}(x, i, j, ..., drop = FALSE)
}
\arguments{
\item{object}{a \code{MinDistPosterior} object}

\item{x}{a \code{MinDistPosterior} object}

\item{i}{an index for subsetting rows}

\item{j}{an index for subsetting columns}

\item{...}{additional arguments passed to subsetting matrices}

\item{drop}{logical -- whether to coerce single-row matrices to vectors}
}
\description{
MinDistPosterior is a \code{GRangesList}-derived container for the
segmentation and maximum a posteriori trio copy number states.
}
\seealso{
\code{\link{denovo}}
}

