% Generated by roxygen2 (4.0.2): do not edit by hand
\name{nMAD}
\alias{nMAD}
\alias{nMAD<-}
\title{Setter and getter for number of median absolute deviations the mean
minimum distance of a genomic interval is from zero}
\usage{
nMAD(object)

nMAD(object) <- value
}
\arguments{
\item{object}{see \code{showMethods("nMAD")}}

\item{value}{a length-one numeric vector}
}
\description{
Setter and getter for number of median absolute deviations the mean
minimum distance of a genomic interval is from zero
}

