% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{ParentOffspringList-class}
\alias{"[",ParentOffspringList,ANY-method}
\alias{"[[",ParentOffspringList,ANY,ANY-method}
\alias{ParentOffspringList}
\alias{ParentOffspringList-class}
\alias{[,ParentOffspringList,ANY-method}
\alias{[[,ParentOffspringList,ANY,ANY-method}
\alias{length,ParentOffspringList-method}
\alias{pedigreeName,ParentOffspringList-method}
\title{A list of \code{ParentOffspring} objects}
\usage{
ParentOffspringList(pedigrees = list())

\S4method{pedigreeName}{ParentOffspringList}(object)

\S4method{[[}{ParentOffspringList,ANY,ANY}(x, i, j, ..., drop = FALSE)

\S4method{[}{ParentOffspringList,ANY}(x, i, j, ..., drop = FALSE)

\S4method{length}{ParentOffspringList}(x)
}
\arguments{
\item{pedigrees}{a list of \code{ParentOffspring} objects}

\item{object}{a \code{ParentOffspringList} object}

\item{x}{a \code{ParentOffspringList} object}

\item{i}{a numeric vector for subsetting the list (optional)}

\item{j}{ignored}

\item{...}{ignored}

\item{drop}{ignored}
}
\description{
Each element of the list is an element of class
\code{\linkS4class{ParentOffspring}}.

Constructor for ParentOffspringList class
}
\section{Slots}{

\describe{
\item{\code{id}}{a character vector of identifiers for the
pedigrees. \code{id} must have the same length as \code{pedigrees}}

\item{\code{pedigrees}}{A list of ParentOffspring objects.}
}}
\examples{
ParentOffspringList()
}

