% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Pedigree}
\alias{Pedigree}
\title{Deprecated function for constructing an instance of class Pedigree}
\usage{
Pedigree(pedigreeInfo, fatherIds = character(), motherIds = character(),
  offspringIds = character())
}
\arguments{
\item{pedigreeInfo}{a \code{data.frame} with column names 'F' (father), 'M' (mother), and 'O' (offspring). Elements of the \code{data.frame} are the sample names.}

\item{fatherIds}{character vector of identifiers for the father}

\item{motherIds}{character vector of identifiers for the mother}

\item{offspringIds}{character vector of identifiers for the offspring}
}
\description{
This function is deprecated and will be removed in a future release.
}
\examples{
Pedigree()
}

