% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pedigreeName}
\alias{pedigreeName}
\title{Accessor for pedigree name}
\usage{
pedigreeName(object)
}
\arguments{
\item{object}{a \code{ParentOffspring} or \code{ParentOffspringList} object}
}
\description{
Accessor for pedigree name
}
\seealso{
\code{\linkS4class{ParentOffspring}} \code{\linkS4class{ParentOffspringList}}
}

