% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pedigreeViewports}
\alias{pedigreeViewports}
\title{Default viewports for plotting log R ratios, BAFs, chromosome
idiogram, and a legend for a case-parent trio}
\usage{
pedigreeViewports()
}
\description{
Default viewports for plotting log R ratios, BAFs, chromosome
idiogram, and a legend for a case-parent trio
}
\examples{
vps <- pedigreeViewports()
}
\seealso{
\code{\link{plotDenovo}} \code{\link{pedigreeGrid}}
}

