% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{plotDenovo}
\alias{plotDenovo}
\alias{plotDenovo,MinDistExperiment,MDRanges-method}
\title{Plot marker-level summaries for a genomic interval of interest}
\usage{
plotDenovo(object, g, param)

\S4method{plotDenovo}{MinDistExperiment,MDRanges}(object, g, param)
}
\arguments{
\item{object}{see \code{showMethods("plotDenovo")}}

\item{g}{a \code{MDRanges} object}

\item{param}{a \code{HmmTrellisParam} object}
}
\description{
Plot marker-level summaries for a genomic interval of interest
}

