% Generated by roxygen2 (4.0.2): do not edit by hand
\name{TrioSet}
\alias{TrioSet}
\title{Deprecated constructor for \code{TrioSet} class}
\usage{
TrioSet(pedigreeData = Pedigree(), sample.sheet, row.names = NULL, lrr, baf,
  featureData, cdfname, drop = TRUE, mindist = NULL, genome = c("hg19",
  "hg18"))
}
\arguments{
\item{pedigreeData}{an object of class \code{Pedigree}}

\item{sample.sheet}{a \code{data.frame} containing metadata on the trios}

\item{row.names}{a character vector providing row identifiers for
the \code{sample.sheet} argument that match the names of the
trios in the \code{pedigreeData} argument.}

\item{lrr}{a matrix of log R ratios}

\item{baf}{a matrix of B allele frequencies}

\item{featureData}{a \code{GenomeAnnotatedDataFrame} object for the SNPs/nonpolymorphic markers}

\item{cdfname}{character string indicating the annotation package used to extract physical position and chromosome of markers}

\item{drop}{logical.  When FALSE, the dimnames on the log R ratio and BAF arrays is set to NULL}

\item{mindist}{can be either NULL or a matrix of the minimum distance}

\item{genome}{character string providing the UCSC genome build}
}
\value{
\code{TrioSet}
}
\description{
The \code{TrioSet} class has been deprecated and may be removed in
a future release.
}
\examples{
path <- system.file("extdata", package="MinimumDistance")
	load(file.path(path, "logRratio.rda"))
	load(file.path(path, "baf.rda"))
	load(file.path(path, "pedigreeInfo.rda"))
	trioSet <- TrioSet(lrr=logRratio,
			   baf=baf,
			   pedigree=Pedigree(pedigreeInfo),
			   cdfname="human610quadv1bCrlmm",
			   genome="hg18")
}

