% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{TrioSetList-class}
\alias{"[",TrioSetList,ANY-method}
\alias{"[[",TrioSetList,ANY,ANY-method}
\alias{$,TrioSetList-method}
\alias{MAP,TrioSetList,GRanges-method}
\alias{TrioSetList-class}
\alias{[,TrioSetList,ANY-method}
\alias{[[,TrioSetList,ANY,ANY-method}
\alias{coerce,TrioSetList,SummarizedExperiment-method}
\alias{length,TrioSetList-method}
\alias{offspringNames,TrioSetList-method}
\alias{pedigree,TrioSetList-method}
\alias{segment2,TrioSetList-method}
\alias{segment2,list-method}
\alias{trios,TrioSetList-method}
\title{Deprecated class for storing low-level genomic data for trios}
\usage{
\S4method{coerce}{TrioSetList,SummarizedExperiment}(from, to)

\S4method{pedigree}{TrioSetList}(object)

\S4method{trios}{TrioSetList}(object)

\S4method{offspringNames}{TrioSetList}(object)

\S4method{[}{TrioSetList,ANY}(x, i, j, ..., drop = FALSE)

\S4method{[[}{TrioSetList,ANY,ANY}(x, i, j, ..., exact = TRUE)

\S4method{length}{TrioSetList}(x)

\S4method{$}{TrioSetList}(x, name)

\S4method{MAP}{TrioSetList,GRanges}(object, ranges, id, TAUP = 1e+10,
  tauMAX = 1 - 5e-08, cnStates = c(-2, -0.4, 0, 0, 0.4, 1),
  pr.nonmendelian = 1.5e-06, mdThr = 0.9, ...)

\S4method{segment2}{TrioSetList}(object, md = NULL, segmentParents = TRUE,
  verbose = TRUE, ...)

\S4method{segment2}{list}(object, pos, chrom, id = NULL, featureNames,
  segmentParents = TRUE, verbose = TRUE, genome, ...)
}
\arguments{
\item{from}{a \code{TrioSetList}}

\item{to}{a \code{SummarizedExperiment}}

\item{object}{a \code{TrioSetList} object}

\item{x}{a \code{TrioSetList}}

\item{i}{a numeric vector for subsetting the chromosomes  (optional)}

\item{j}{a numeric vector for subsetting trios (optional)}

\item{...}{additional arguments passed to subsetting methods for matrices and data frames}

\item{drop}{logical. Whether to simplify matrices to numeric
vectors.  This should be left as FALSE.}

\item{exact}{ignored}

\item{name}{character string of a variable name in the phenoData}

\item{ranges}{a \code{GRanges} object}

\item{id}{a character vector of trio identifiers}

\item{TAUP}{length-one numeric vector.  Larger values decrease the
probability of transitioning to an different state.}

\item{tauMAX}{the maximum allowed transition probability}

\item{cnStates}{a length-six numeric vector profiving initial
values for the mean copy number for each of the 6 states}

\item{pr.nonmendelian}{a length-one numeric vector indicating the
probability of a non-Mendelian copy number alteration in the offspring}

\item{mdThr}{a length-one numeric vector indicating the minimum
value of the mean minimum distance. Segments with absolute mean
value less than \code{mdThr} are not called.}

\item{md}{a list of minimum distance matrices. Length of list
should be the same as the length of the \code{TrioSetList} object.}

\item{segmentParents}{logical. Whether to segment the parental log R ratios.}

\item{verbose}{logical. Whether to display messages indicating progress.}

\item{pos}{a list of the genomic positions (integers)}

\item{chrom}{list of chromosome names}

\item{featureNames}{a list of the marker names}

\item{genome}{a character vector indicating the UCSC genome build
used for the annotation (i.e., 'hg18' or 'hg19').}
}
\description{
This class is deprecated and will be defunct in a future release.
}
\section{Slots}{

\describe{
\item{\code{fatherPhenoData}}{\code{AnnotatedDataFrame} containing covariates for the father}

\item{\code{motherPhenoData}}{\code{AnnotatedDataFrame} containing covariates for the mother}

\item{\code{pedigree}}{an object of class \code{Pedigree}}
}}

