% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{FilterParamMD-class}
\alias{FilterParamMD-class}
\title{A class for filtering genomic intervals called by MinimumDistance}
\description{
Options for filtering include the number of markers spanned by a
segment, the posterior probability of the maximum a posteriori
estimate of the trio copy number state, and the trio copy number
state.
}
\seealso{
denovo
}

