% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{MDRanges-class}
\alias{MDRanges}
\alias{MDRanges-class}
\title{A \code{GRanges}-derived class}
\usage{
MDRanges(..., posteriors)
}
\arguments{
\item{...}{additional arguments to \code{GRanges} constructor}

\item{posteriors}{a \code{DataFrame}}
}
\description{
Contains maximum a posteriori estimates for each genomic interval
}
\examples{
MDRanges()
}

