% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{Pedigree-class}
\alias{"[",Pedigree,ANY-method}
\alias{Pedigree-class}
\alias{[,Pedigree,ANY-method}
\alias{dim,Pedigree-method}
\alias{offspringNames,Pedigree-method}
\alias{trios,Pedigree-method}
\title{Deprecated class for storing pedigree data}
\usage{
\S4method{trios}{Pedigree}(object)

\S4method{offspringNames}{Pedigree}(object)

\S4method{[}{Pedigree,ANY}(x, i, j, ..., drop = FALSE)

\S4method{dim}{Pedigree}(x)
}
\arguments{
\item{object}{a \code{Pedigree} object}

\item{x}{a \code{Pedigree} object}

\item{i}{a numeric vector for subsetting  (optional)}

\item{j}{ignored}

\item{...}{ignored}

\item{drop}{ignored}
}
\description{
Deprecated class for storing pedigree data
}
\section{Slots}{

\describe{
\item{\code{trios}}{a \code{data.frame} with colnames 'F', 'M', and 'O'
containing sample identifiers for the father (F), mother (M), and
offspring (O).}

\item{\code{trioIndex}}{a \code{data.frame}}
}}

